requirements.txt
Django>=4.2,<6.0
psycopg2-binary>=2.9
python-decouple>=3.8
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
channels>=4.0
channels-redis>=4.1
daphne>=4.0
openai-whisper>=20231117
numpy>=1.24
torch>=2.0
hypothesis>=6.100
pytest>=8.0
pytest-django>=4.8
pytest-asyncio>=0.23
